Refactor BigModel to be more flexible in its implementation
